+++ src_base/xeve_util.c
@@ -2657,7 +2657,7 @@ int xeve_pic_finish(XEVE_CTX *ctx, XEVE_BITB *bitb, XEVE_STAT *stat)
 
         xeve_bsw_deinit(bs);
         stat->sei_size = (int)(bs->cur - cur_tmp);
-        xeve_eco_nalu_len_update(size_field, stat->sei_size - 4);
+        xeve_eco_nal_unit_len(size_field, stat->sei_size - 4);
     }
 
     /* expand current encoding picture, if needs */
@@ -3547,351 +3547,6 @@ void xeve_recon(XEVE_CTX * ctx, XEVE_CORE * core, s16 *coef, pel *pred, int is_c
     xeve_recon_blk(coef, pred, is_coef, cuw, cuh, s_rec, rec, bit_depth);
 }
 
-int xeve_enc(XEVE_CTX * ctx, XEVE_BITB * bitb, XEVE_STAT * stat)
-{
-    int            ret;
-    int            gop_size, pic_cnt;
-
-    pic_cnt = ctx->pic_icnt - ctx->frm_rnum;
-    gop_size = ctx->param.gop_size;
-
-    ctx->force_slice = ((ctx->pic_ticnt % gop_size >= ctx->pic_ticnt - pic_cnt + 1) && FORCE_OUT(ctx)) ? 1 : 0;
-
-    xeve_assert_rv(bitb->addr && bitb->bsize > 0, XEVE_ERR_INVALID_ARGUMENT);
-
-    /* initialize variables for a picture encoding */
-    ret = ctx->fn_enc_pic_prepare(ctx, bitb, stat);
-    xeve_assert_rv(ret == XEVE_OK, ret);
-
-    /* encode parameter set */
-    ret = ctx->fn_enc_header(ctx);
-    xeve_assert_rv(ret == XEVE_OK, ret);
-
-    /* encode one picture */
-    ret = ctx->fn_enc_pic(ctx, bitb, stat);
-    xeve_assert_rv(ret == XEVE_OK, ret);
-
-    /* finishing of encoding a picture */
-    ctx->fn_enc_pic_finish(ctx, bitb, stat);
-    xeve_assert_rv(ret == XEVE_OK, ret);
-
-    return XEVE_OK;
-}
-
-
-int xeve_push_frm(XEVE_CTX * ctx, XEVE_IMGB * img)
-{
-    XEVE_PIC  * pic, * spic;
-    XEVE_PICO * pico;
-    XEVE_IMGB * imgb;
-
-    int ret;
-
-    ret = ctx->fn_get_inbuf(ctx, &imgb);
-    xeve_assert_rv(XEVE_OK == ret, ret);
-
-    imgb->cs = ctx->param.cs;
-    xeve_imgb_cpy(imgb, img);
-
-    if (ctx->fn_pic_flt != NULL)
-    {
-        ctx->fn_pic_flt(ctx, imgb);
-    }
-
-    ctx->pic_icnt++;
-    ctx->pico_idx = ctx->pic_icnt % ctx->pico_max_cnt;
-    pico = ctx->pico_buf[ctx->pico_idx];
-    pico->pic_icnt = ctx->pic_icnt;
-    pico->is_used = 1;
-    pic = &pico->pic;
-    ctx->pico = pico;
-
-    PIC_ORIG(ctx) = pic;
-
-    /* set pushed image to current input (original) image */
-    xeve_mset(pic, 0, sizeof(XEVE_PIC));
-
-    pic->buf_y = imgb->baddr[0];
-    pic->buf_u = imgb->baddr[1];
-    pic->buf_v = imgb->baddr[2];
-
-    pic->y = imgb->a[0];
-    pic->u = imgb->a[1];
-    pic->v = imgb->a[2];
-
-    pic->w_l = imgb->w[0];
-    pic->h_l = imgb->h[0];
-    pic->w_c = imgb->w[1];
-    pic->h_c = imgb->h[1];
-
-    pic->s_l = STRIDE_IMGB2PIC(imgb->s[0]);
-    pic->s_c = STRIDE_IMGB2PIC(imgb->s[1]);
-
-    pic->imgb = imgb;
-    /* generate sub-picture for RC and Forecast */
-    if (ctx->param.use_fcst)
-    {
-        spic = pico->spic;
-        xeve_gen_subpic(pic->y, spic->y, spic->w_l, spic->h_l, pic->s_l, spic->s_l, 10);
-
-        xeve_mset(pico->sinfo.map_pdir, 0, sizeof(u8) * ctx->fcst.f_blk);
-        xeve_mset(pico->sinfo.map_pdir_bi, 0, sizeof(u8) * ctx->fcst.f_blk);
-        xeve_mset(pico->sinfo.map_mv, 0, sizeof(s16) * ctx->fcst.f_blk * REFP_NUM * MV_D);
-        xeve_mset(pico->sinfo.map_mv_bi, 0, sizeof(s16) * ctx->fcst.f_blk * REFP_NUM * MV_D);
-        xeve_mset(pico->sinfo.map_mv_pga, 0, sizeof(s16) * ctx->fcst.f_blk * REFP_NUM * MV_D);
-        xeve_mset(pico->sinfo.map_uni_lcost, 0, sizeof(s32) * ctx->fcst.f_blk * 4);
-        xeve_mset(pico->sinfo.map_bi_lcost, 0, sizeof(s32) * ctx->fcst.f_blk);
-        xeve_mset(pico->sinfo.map_qp_blk, 0, sizeof(s32) * ctx->fcst.f_blk);
-        xeve_mset(pico->sinfo.map_qp_scu, 0, sizeof(s8) * ctx->f_scu);
-        xeve_mset(pico->sinfo.transfer_cost, 0, sizeof(u16) * ctx->fcst.f_blk);
-        xeve_picbuf_expand(spic, spic->pad_l, spic->pad_c, ctx->sps.chroma_format_idc);
-    }
-    return XEVE_OK;
-}
-
-
-void xeve_platform_init_func(XEVE_CTX * ctx)
-{
-#if X86_SSE
-    int check_cpu, support_sse, support_avx, support_avx2;
-
-    check_cpu = xeve_check_cpu_info();
-    support_sse  = (check_cpu >> 1) & 1;
-    support_avx  = check_cpu & 1;
-    support_avx2 = (check_cpu >> 2) & 1;
-
-    if (support_avx2)
-    {
-        xeve_func_sad               = xeve_tbl_sad_16b_avx;
-        xeve_func_ssd               = xeve_tbl_ssd_16b_sse;
-        xeve_func_diff              = xeve_tbl_diff_16b_sse;
-        xeve_func_satd              = xeve_tbl_satd_16b_sse;
-        xeve_func_mc_l              = xeve_tbl_mc_l_avx;
-        xeve_func_mc_c              = xeve_tbl_mc_c_avx;
-        xeve_func_average_no_clip   = &xeve_average_16b_no_clip_sse;
-        ctx->fn_itxb                = &xeve_tbl_itxb_avx;
-        xeve_func_txb               = &xeve_tbl_txb_avx;
-    }
-    else if (support_sse)
-    {
-        xeve_func_sad               = xeve_tbl_sad_16b_sse;
-        xeve_func_ssd               = xeve_tbl_ssd_16b_sse;
-        xeve_func_diff              = xeve_tbl_diff_16b_sse;
-        xeve_func_satd              = xeve_tbl_satd_16b_sse;
-        xeve_func_mc_l              = xeve_tbl_mc_l_sse;
-        xeve_func_mc_c              = xeve_tbl_mc_c_sse;
-        xeve_func_average_no_clip   = &xeve_average_16b_no_clip_sse;
-        ctx->fn_itxb                = &xeve_tbl_itxb_sse;
-        xeve_func_txb               = &xeve_tbl_txb; /*to be updated*/
-    }
-    else
-#endif
-    {
-        xeve_func_sad               = xeve_tbl_sad_16b;
-        xeve_func_ssd               = xeve_tbl_ssd_16b;
-        xeve_func_diff              = xeve_tbl_diff_16b;
-        xeve_func_satd              = xeve_tbl_satd_16b;
-        xeve_func_mc_l              = xeve_tbl_mc_l;
-        xeve_func_mc_c              = xeve_tbl_mc_c;
-        xeve_func_average_no_clip   = &xeve_average_16b_no_clip;
-        ctx->fn_itxb                = &xeve_tbl_itxb;
-        xeve_func_txb               = &xeve_tbl_txb;
-    }
-}
-
-int xeve_platform_init(XEVE_CTX * ctx)
-{
-    int ret = XEVE_ERR_UNKNOWN;
-
-    /* create mode decision */
-    ret = xeve_mode_create(ctx, 0);
-    xeve_assert_rv(XEVE_OK == ret, ret);
-
-    /* create intra prediction analyzer */
-    ret = xeve_pintra_create(ctx, 0);
-    xeve_assert_rv(XEVE_OK == ret, ret);
-
-    /* create inter prediction analyzer */
-    if (ctx->param.profile == XEVE_PROFILE_BASELINE)
-    {
-        ret = xeve_pinter_create(ctx, 0);
-        xeve_assert_rv(XEVE_OK == ret, ret);
-    }
-
-    ctx->fn_ready             = xeve_ready;
-    ctx->fn_flush             = xeve_flush;
-    ctx->fn_enc               = xeve_enc;
-    ctx->fn_enc_header        = xeve_header;
-    ctx->fn_enc_pic           = xeve_pic;
-    ctx->fn_enc_pic_prepare   = xeve_pic_prepare;
-    ctx->fn_enc_pic_finish    = xeve_pic_finish;
-    ctx->fn_push              = xeve_push_frm;
-    ctx->fn_deblock           = xeve_deblock;
-    ctx->fn_picbuf_expand     = xeve_pic_expand;
-    ctx->fn_get_inbuf         = xeve_picbuf_get_inbuf;
-    ctx->fn_loop_filter       = xeve_loop_filter;
-    ctx->fn_encode_sps        = xeve_encode_sps;
-    ctx->fn_encode_pps        = xeve_encode_pps;
-    ctx->fn_eco_sh            = xeve_eco_sh;
-    ctx->fn_eco_split_mode    = xeve_eco_split_mode;
-    ctx->fn_eco_sbac_reset    = xeve_sbac_reset;
-    ctx->fn_eco_coef          = xeve_eco_coef;
-    ctx->fn_eco_pic_signature = xeve_eco_pic_signature;
-    ctx->fn_tq                = xeve_sub_block_tq;
-    ctx->fn_rdoq_set_ctx_cc   = xeve_rdoq_set_ctx_cc;
-    ctx->fn_itdp              = xeve_itdq;
-    ctx->fn_recon             = xeve_recon;
-    ctx->fn_deblock_tree      = xeve_deblock_tree;
-    ctx->fn_deblock_unit      = xeve_deblock_unit;
-    ctx->fn_set_tile_info     = xeve_set_tile_info;
-    ctx->fn_rdo_intra_ext     = NULL;
-    ctx->fn_rdo_intra_ext_c   = NULL;
-    ctx->pic_dbk              = NULL;
-    ctx->fn_pocs              = NULL;
-    ctx->fn_pic_flt           = NULL;
-    ctx->pf                   = NULL;
-
-    xeve_platform_init_func(ctx);
-
-    return XEVE_OK;
-}
-
-void xeve_platform_deinit(XEVE_CTX * ctx)
-{
-    xeve_assert(ctx->pf == NULL);
-
-    ctx->fn_ready = NULL;
-    ctx->fn_flush = NULL;
-    ctx->fn_enc = NULL;
-    ctx->fn_enc_pic = NULL;
-    ctx->fn_enc_pic_prepare = NULL;
-    ctx->fn_enc_pic_finish = NULL;
-    ctx->fn_push = NULL;
-    ctx->fn_deblock = NULL;
-    ctx->fn_picbuf_expand = NULL;
-    ctx->fn_get_inbuf = NULL;
-}
-
-int xeve_create_bs_buf(XEVE_CTX  * ctx, int max_bs_buf_size)
-{
-    u8 * bs_buf, *bs_buf_temp;
-    if (ctx->param.threads > 1)
-    {
-        bs_buf = (u8 *)xeve_malloc(sizeof(u8 *) * (ctx->param.threads - 1) * max_bs_buf_size);
-        for (int task_id = 1; task_id < ctx->param.threads; task_id++)
-        {
-            bs_buf_temp = bs_buf + ((task_id - 1) * max_bs_buf_size);
-            xeve_bsw_init(&ctx->bs[task_id], bs_buf_temp, max_bs_buf_size, NULL);
-            ctx->bs[task_id].pdata[1] = &ctx->sbac_enc[task_id];
-        }
-    }
-    return XEVE_OK;
-}
-
-int xeve_delete_bs_buf(XEVE_CTX  * ctx)
-{
-    if (ctx->param.threads > 1)
-    {
-        u8 * bs_buf_temp = ctx->bs[1].beg;
-        if (bs_buf_temp != NULL)
-        {
-            xeve_mfree(bs_buf_temp);
-        }
-        bs_buf_temp = NULL;
-    }
-    return XEVE_OK;
-}
-
-
-int xeve_encode_sps(XEVE_CTX * ctx)
-{
-    XEVE_BSW * bs = &ctx->bs[0];
-    XEVE_SPS * sps = &ctx->sps;
-    XEVE_NALU  nalu;
-
-    u8* size_field = bs->cur;
-    u8* cur_tmp = bs->cur;
-
-    /* nalu header */
-    xeve_set_nalu(&nalu, XEVE_SPS_NUT, 0);
-    xeve_eco_nalu(bs, &nalu);
-
-    /* sequence parameter set*/
-    xeve_set_sps(ctx, &ctx->sps);
-    xeve_assert_rv(xeve_eco_sps(bs, sps) == XEVE_OK, XEVE_ERR_INVALID_ARGUMENT);
-
-    /* de-init BSW */
-    xeve_bsw_deinit(bs);
-
-    /* write the bitstream size */
-    xeve_eco_nalu_len_update(size_field, (int)(bs->cur - cur_tmp) - 4);
-
-    return XEVE_OK;
-}
-
-int xeve_encode_pps(XEVE_CTX * ctx)
-{
-    XEVE_BSW * bs = &ctx->bs[0];
-    XEVE_SPS * sps = &ctx->sps;
-    XEVE_PPS * pps = &ctx->pps;
-    XEVE_NALU  nalu;
-    u8       * size_field = bs->cur;
-    u8       * cur_tmp = bs->cur;
-
-    /* nalu header */
-    xeve_set_nalu(&nalu, XEVE_PPS_NUT, ctx->nalu.nuh_temporal_id);
-    xeve_eco_nalu(bs, &nalu);
-
-    /* sequence parameter set*/
-    xeve_set_pps(ctx, &ctx->pps);
-    xeve_assert_rv(xeve_eco_pps(bs, sps, pps) == XEVE_OK, XEVE_ERR_INVALID_ARGUMENT);
-
-    /* de-init BSW */
-    xeve_bsw_deinit(bs);
-
-    /* write the bitstream size */
-    xeve_eco_nalu_len_update(size_field, (int)(bs->cur - cur_tmp) - 4);
-    return XEVE_OK;
-}
-
-int xeve_check_frame_delay(XEVE_CTX * ctx)
-{
-    if (ctx->pic_icnt < ctx->frm_rnum)
-    {
-        return XEVE_OK_OUT_NOT_AVAILABLE;
-    }
-    return XEVE_OK;
-}
-
-int xeve_check_more_frames(XEVE_CTX * ctx)
-{
-    XEVE_PICO  * pico;
-    int           i;
-
-    if(FORCE_OUT(ctx))
-    {
-        /* pseudo xeve_push() for bumping process ****************/
-        ctx->pic_icnt++;
-        /**********************************************************/
-
-        for(i=0; i<ctx->pico_max_cnt; i++)
-        {
-            pico = ctx->pico_buf[i];
-            if(pico != NULL)
-            {
-                if(pico->is_used == 1)
-                {
-                    return XEVE_OK;
-                }
-            }
-        }
-
-        return XEVE_OK_NO_MORE_FRM;
-    }
-
-    return XEVE_OK;
-}
-
 int xeve_malloc_1d(void** dst, int size)
 {
     int ret;
@@ -4201,134 +3856,3 @@ void xeve_set_tile_in_slice(XEVE_CTX * ctx)
     }
 }
 
-int xeve_param_init(XEVE_PARAM* param)
-{
-    xeve_mset(param, 0, sizeof(XEVE_PARAM));
-
-    param->profile                    = XEVE_PROFILE_BASELINE;
-    param->qp                         = 32;
-    param->crf                        = 32;
-    param->bframes                    = 15;
-    param->codec_bit_depth            = 10;
-    param->lookahead                  = 17;
-    param->use_deblock                = 1;
-    param->threads                    = 1;
-    param->rdo_dbk_switch             = 1;
-    param->tile_rows                  = 1;
-    param->tile_columns               = 1;
-    param->num_slice_in_pic           = 1;
-
-    return XEVE_OK;
-}
-
-int xeve_param_apply_ppt_baseline(XEVE_PARAM* param, int profile, int preset, int tune)
-{
-    if (profile != XEVE_PROFILE_BASELINE)
-    {
-        return XEVE_ERR;
-    }
-
-    param->profile = XEVE_PROFILE_BASELINE;
-
-    if (preset == XEVE_PRESET_FAST)
-    {
-        param->max_cu_intra      = 32;
-        param->min_cu_intra      = 4;
-        param->max_cu_inter      = 64;
-        param->min_cu_inter      = 8;
-        param->me_ref_num        = 1;
-        param->me_algo           = 1;
-        param->me_range          = 32;
-        param->me_sub            = 2;
-        param->me_sub_pos        = 2;
-        param->me_sub_range      = 1;
-        param->skip_th           = 0;
-        param->merge_num         = 2;
-        param->rdoq              = 1;
-        param->cabac_refine      = 1;
-        param->rdo_dbk_switch    = 0;
-    }
-    else if (preset == XEVE_PRESET_MEDIUM)
-    {
-        param->max_cu_intra      = 32;
-        param->min_cu_intra      = 4;
-        param->max_cu_inter      = 64;
-        param->min_cu_inter      = 8;
-        param->me_ref_num        = 1;
-        param->me_algo           = 1;
-        param->me_range          = 64;
-        param->me_sub            = 2;
-        param->me_sub_pos        = 4;
-        param->me_sub_range      = 1;
-        param->skip_th           = 0;
-        param->merge_num         = 3;
-        param->rdoq              = 1;
-        param->cabac_refine      = 1;
-        param->rdo_dbk_switch    = 0;
-    }
-    else if (preset == XEVE_PRESET_SLOW)
-    {
-        param->max_cu_intra      = 32;
-        param->min_cu_intra      = 4;
-        param->max_cu_inter      = 64;
-        param->min_cu_inter      = 8;
-        param->me_ref_num        = 1;
-        param->me_algo           = 1;
-        param->me_range          = 128;
-        param->me_sub            = 3;
-        param->me_sub_pos        = 4;
-        param->me_sub_range      = 2;
-        param->skip_th           = 0;
-        param->merge_num         = 3;
-        param->rdoq              = 1;
-        param->cabac_refine      = 1;
-        param->rdo_dbk_switch    = 1;
-    }
-    else if (preset == XEVE_PRESET_PLACEBO)
-    {
-        param->max_cu_intra      = 64;
-        param->min_cu_intra      = 4;
-        param->max_cu_inter      = 64;
-        param->min_cu_inter      = 4;
-        param->me_ref_num        = 2;
-        param->me_algo           = 2;
-        param->me_range          = 384;
-        param->me_sub            = 3;
-        param->me_sub_pos        = 8;
-        param->me_sub_range      = 3;
-        param->skip_th           = 0;
-        param->merge_num         = 4;
-        param->rdoq              = 1;
-        param->cabac_refine      = 1;
-        param->rdo_dbk_switch    = 1;
-    }
-    else
-    {
-        return XEVE_ERR;
-    }
-
-    if (tune != XEVE_TUNE_NONE)
-    {
-        if (tune == XEVE_TUNE_ZEROLATENCY)
-        {
-            param->aq_mode = 1;
-            param->lookahead = 0;
-            param->cutree = 0;
-            param->bframes = 0;
-            param->me_ref_num = 1;
-            param->ref_pic_gap_length = 1;
-            param->use_fcst = 1;
-            param->inter_slice_type = 1;
-        }
-        else if (tune == XEVE_TUNE_PSNR)
-        {
-            param->aq_mode = 0;
-        }
-        else
-        {
-            return XEVE_ERR;
-        }
-    }
-
-    return XEVE_OK;
-}
