+++ src_main/xevem_util.c
@@ -4011,6 +4011,7 @@ int xevem_platform_init(XEVE_CTX * ctx)
 
     ctx->fn_ready           = xevem_ready;
     ctx->fn_flush           = xevem_flush;
+    ctx->fn_enc_pic         = xevem_pic;
     ctx->fn_enc_header      = xevem_header;
     ctx->fn_enc_pic_prepare = xevem_pic_prepare;
     ctx->fn_loop_filter     = xevem_loop_filter;
@@ -4068,7 +4069,7 @@ int xevem_encode_sps(XEVE_CTX * ctx)
     xeve_bsw_deinit(bs);
 
     /* write the bitstream size */
-    xeve_eco_nalu_len_update(size_field, (int)(bs->cur - cur_tmp) - 4);
+    xeve_eco_nal_unit_len(size_field, (int)(bs->cur - cur_tmp) - 4);
 
     return XEVE_OK;
 }
@@ -4088,7 +4089,7 @@ int xevem_encode_aps(XEVE_CTX * ctx, XEVE_APS_GEN * aps)
     xeve_assert_rv(xevem_eco_aps_gen(bs, aps, ctx->sps.bit_depth_luma_minus8 + 8) == XEVE_OK, XEVE_ERR_INVALID_ARGUMENT);
 
     xeve_bsw_deinit(bs);
-    xeve_eco_nalu_len_update(size_field, (int)(bs->cur - cur_tmp) - 4);
+    xeve_eco_nal_unit_len(size_field, (int)(bs->cur - cur_tmp) - 4);
 
     return XEVE_OK;
 }
@@ -4113,7 +4114,7 @@ int xevem_encode_pps(XEVE_CTX * ctx)
     xeve_bsw_deinit(bs);
 
     /* write the bitstream size */
-    xeve_eco_nalu_len_update(size_field, (int)(bs->cur - cur_tmp) - 4);
+    xeve_eco_nal_unit_len(size_field, (int)(bs->cur - cur_tmp) - 4);
 
     return XEVE_OK;
 }
